Input validation logic was added for RESTful services.

+++ src/main/java/com/h2rd/refactoring/web/UserResource.java
@@ -14,6 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static org.apache.commons.lang3.StringUtils.isBlank;
 
 @Path("/users")
 @Component
@@ -21,6 +22,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 public class UserResource {
 
     private static final String USER_NOT_FOUND = "User not found.";
+    private static final String USER_ALREADY_EXISTS = "User already exists.";
 
     @Autowired
     private UserOperations userDao;
@@ -33,33 +35,75 @@ public class UserResource {
     @POST
     @Path("/add")
     public Response addUser(final User user) {
-        validateUser(user);
-        userDao.saveUser(user);
-        return Response.ok().entity(user).build();
+        final List<String> errors = validateUser(user,true,true,true);
+        if (!errors.isEmpty()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(errors.get(0)).build();
+        }
+
+        final User existingUser = userDao.findUserById(user.getEmail());
+        if (existingUser == null) {
+            userDao.saveUser(user);
+            return Response.ok().entity(user).build();
+        } else {
+            return Response.status(Response.Status.BAD_REQUEST).entity(USER_ALREADY_EXISTS).build();
+        }
     }
 
-    private void validateUser(final User user) {
+    private List<String> validateUser(final User user, final boolean checkName, final boolean checkEmail, final boolean checkRoles) {
+        final List<String> errors = new ArrayList<>();
+        if (user == null) {
+            errors.add("User is mandatory!");
+        } else {
+            if (checkEmail && isBlank(user.getEmail())) {
+                errors.add("Email is mandatory!");
+            }
 
+            if (checkName && isBlank(user.getName())) {
+                errors.add("Name is mandatory!");
+            }
+
+            if (checkRoles && (user.getRoles() == null || user.getRoles().isEmpty())) {
+                errors.add("User has to have at least one role!");
+            }
+        }
+
+        return errors;
     }
 
     @PUT
     @Path("/update")
     public Response updateUser(final User user) {
-        //validateRoles(user);
-        userDao.updateUser(user);
-        return Response.ok().entity(user).build();
+        final List<String> errors = validateUser(user,true,true,true);
+
+        if (!errors.isEmpty()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(errors.get(0)).build();
+        }
+
+        final User existingUser = userDao.findUserById(user.getEmail());
+        if (existingUser != null) {
+            userDao.updateUser(user);
+            return Response.ok().entity(user).build();
+        } else {
+            return Response.status(Response.Status.NOT_FOUND).entity(USER_NOT_FOUND).build();
+        }
     }
 
     @DELETE
     @Path("/delete")
     public Response deleteUser(@QueryParam("email") final String email) {
-
         final User user = new User("", email, new ArrayList<>());
-        //validateEmail(email)
+        final List<String> errors = validateUser(user,false,true,false);
+        if (!errors.isEmpty()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(errors.get(0)).build();
+        }
 
-        userDao.deleteUser(user);
-
-        return Response.ok().entity(user).build();
+        final User existingUser = userDao.findUserById(user.getEmail());
+        if (existingUser != null) {
+            userDao.deleteUser(user);
+            return Response.ok().entity(user).build();
+        } else {
+            return Response.status(Response.Status.NOT_FOUND).entity(USER_NOT_FOUND).build();
+        }
     }
 
     @GET
@@ -79,8 +123,6 @@ public class UserResource {
     @Path("/search")
     @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
     public Response findUser(@QueryParam("name") final String name) {
-        //validateEmail(name)
-
         final User user = userDao.findUser(name);
         if (user != null) {
             return Response.ok().entity(user).build();
